Deleted EditCompetencyModelModal copy
